requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.4
future==0.18.2
gunicorn==20.0.4
idna==2.9
line-bot-sdk==1.8.0
lxml==4.5.0
psycopg2==2.8.4
pytz==2019.3
requests==2.23.0
soupsieve==2.0
sqlparse==0.3.1
static3==0.7.0
urllib3==1.25.8
